Fix remove account confirmation
