+++ ui/components/multichain/account-list-item/account-list-item.js
@@ -107,7 +107,7 @@ export const AccountListItem = ({
         // Without this check, the account will be selected after
         // the account options menu closes
         if (!accountOptionsMenuOpen) {
-          onClick();
+          onClick?.();
         }
       }}
     >
@@ -153,7 +153,7 @@ export const AccountListItem = ({
                 as="button"
                 onClick={(e) => {
                   e.stopPropagation();
-                  onClick();
+                  onClick?.();
                 }}
                 variant={TextVariant.bodyMdMedium}
                 className="multichain-account-list-item__account-name__button"
@@ -311,7 +311,7 @@ AccountListItem.propTypes = {
   /**
    * Function to execute when the item is clicked
    */
-  onClick: PropTypes.func.isRequired,
+  onClick: PropTypes.func,
   /**
    * Function that closes the menu
    */
