+++ ui/pages/pages.scss
@@ -27,6 +27,8 @@
 @import 'settings/index';
 @import 'snaps/snaps-list/index';
 @import 'snaps/snap-view/index';
+@import 'create-snap-account/index';
+@import 'remove-snap-account/index';
 @import 'swaps/index';
 @import 'token-allowance/index';
 @import 'token-details/index';
