--- ui/pages/remove-snap-account/components/index.ts
@@ -1,1 +1,0 @@
-export { default as RemoveSnapAccountContent } from './remove-snap-account-content';
